joining all helper functions to utils.py file
